cli: share wizard between run and config commands
